requirements.txt
flask
line-bot-sdk
PyTelegramBotAPI

